refactor: Track Appointment Meeting Join
